Extract UpdateMetadataForIndexCallable
Extract UpdateMetadataForIndexCallable to be a standalone class.
